+++ components/Tabs/vue/Tabs.vue
@@ -14,6 +14,10 @@ const props = withDefaults(
      */
     tabs: TabConfig[]
     /**
+     * The active tab
+     */
+    activeId?: string
+    /**
      * Appearance of tabs
      */
     variant?: keyof typeof variants
@@ -32,9 +36,17 @@ const emit = defineEmits<{
    * @param tab new tab selected
    */
   (event: 'switch', tab: TabConfig): void
+  (event: 'update:activeId', tab: string): void
 }>()
 
-const activeId = ref(props.tabs.find((tab) => tab.active)?.id)
+const activeId = ref(props.activeId ?? props.tabs.find((tab) => tab.active)?.id)
+
+watch(
+  () => props.activeId,
+  (id) => {
+    activeId.value = id
+  }
+)
 
 const activeMarkerStyle = ref<
   { left?: string; width?: string; transitionProperty?: string } | undefined
@@ -95,6 +107,7 @@ function navigate(shift: number) {
       : shiftedIndex
   activeId.value = props.tabs[nextIndex].id
   $tab.value?.[nextIndex]?.$el.focus()
+  emit('update:activeId', props.tabs[nextIndex].id)
   emit('switch', props.tabs[nextIndex])
 }