+++ css/src/tw-icon-extractor.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { getHtmlAttributes } from './tw-icon-extractor'
+import { getHtmlAttributes, IconExtractor } from './tw-icon-extractor'
 
 describe('getHtmlAttributes', () => {
   it('should return null if no attributes are found', () => {
@@ -18,5 +18,30 @@ describe('getHtmlAttributes', () => {
       names: ['fill'],
       values: ['red'],
     })
+  })
+
+  it('should take care of isolated lines with curlies', () => {
+    expect(getHtmlAttributes("fill={'red'}")).toEqual({
+      names: ['fill'],
+      values: ['red'],
+    })
+  })
+})
+
+describe('IconExtractor', () => {
+  it('should extract colors', () => {
+    const ext = IconExtractor(
+      '<wind-keep fillColor="red-300" strokeColor="indigo-800" />'
+    )
+    expect(ext).toContain('icon-light-red-300')
+    expect(ext).toContain('icon-dark-indigo-800')
+  })
+
+  it('should extract colors in curlies', () => {
+    const ext = IconExtractor(
+      "<wind-keep fillColor={'red-300'} strokeColor={'indigo-800'} />"
+    )
+    expect(ext).toContain('icon-light-red-300')
+    expect(ext).toContain('icon-dark-indigo-800')
   })
 })